[FIX] Format partner name properly when searching from views
Currently, when searching for a partner with the context variable `{'show_address': True}` combined `{'show_vat': True}` or `{'show_email': True}`, it does not show either VAT or email in the results before clicking on the partner.
This is because it shows only the first line that returns `name_get` function (`_get_name`), and having in the context `{'show_address': True}` formats it with a carriage return just after the customer's name, leaving the VAT and the email at the end.

Since the VAT and email fields are being searched by default in the _name_search function, it is logical and feasible that they are displayed in the first line as if they were search matches.
---
Example of how it is now displayed with context `{'show_address': True, 'show_vat': True, 'show_email': True}`:
Azure Interior
4557 De Silva St
Fremont CA 94538
United States <azure.Interior24@example.com> - ESR2537651H

Example of how it would be displayed after changes with context `{'show_address': True, 'show_vat': True, 'show_email': True}`:
Customer Azure Interior <azure.Interior24@example.com> - ESR2537651H
4557 De Silva St
Fremont CA 94538
United States
---
In this way, you can see the field you are looking for directly before selecting the record.

This changes are proposed because on Purchase Order, you can view the VAT next to the partner. 
On Sales Order and Account Move (invoice) you can't because `{'show_address': True}` is present on the view field context.